improve ui and fix some logic to display

+++ app/Livewire/Admin/Job/JobEdit.php
@@ -120,10 +120,15 @@ class JobEdit extends Component
 
     public function deleteAddress(string|int $id): void
     {
-        Address::findOrFail($id)->delete();
-        $this->addresses = $this->job->addresses()->get();
-        $this->alert('success', trans('Delete success'));
-        $this->dispatch('refresh');
+        if ($this->job->addresses->count() > 1) {
+            Address::findOrFail($id)->delete();
+            $this->addresses = $this->job->addresses()->get();
+            $this->alert('success', trans('Delete success'));
+            $this->dispatch('refresh');
+            return;
+        }
+
+        $this->alert('warning', trans('Can not delete all addresses'));
     }
 
     #[On('refresh')]
